Refactor: /utils/apiScripts.js -> /api/apiScripts.js
