+++ src/pages/Activity.jsx
@@ -5,14 +5,14 @@ import { PrevButton, NextButton } from "../components/PageButton";
 import RateForm from "../components/RateForm";
 import ContentBoard from "../components/ContentBoard";
 import ActivityBar from "../components/ActivityBar";
-import { getActivities, editActivityById } from "../utils/activity";
+import { getActivities, editActivityById } from "../api/activity";
 
 const CONTENT_BOARD_PIXEL_WIDTH = 400;
 const CONTENT_BOARD_PIXEL_HEIGHT = 150;
 
 const Title = styled.p`
   margin: 10px 0px 10px 30px;
-  color: ${({ theme }) => theme.mintColors.mainMint};
+  color: ${({ theme }) => theme.background.sub};
   text-align: left;
   font-size: 50px;
 `;
